requirements.txt
# FlowSentrix - Python Dependencies
# Install: pip install -r requirements.txt

# Core capture engine
scapy>=2.5.0

# CLI interface
rich>=13.0.0

# Web GUI backend
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Config parsing
pyyaml>=6.0
